+++ release_notes.rst
@@ -1,9 +1,21 @@
 Release Notes
 =============
 
+v2.0.3
+------
+* Merging v2 hotfixes
+
 v2.0.1
 ------
 * Fix for handling single failures in a batch of outgoing emails
+
+v2.0.0.2
+------
+* Atomic decorator on event fetching
+
+v2.0.0.1
+------
+* Fix unique constraint when bulk creating emails
 
 v2.0.0
 ------